+++ main/utils.h
@@ -18,11 +18,11 @@
             *(a + q + 1) = *(a + q); }}}
 
 /** first 8 bits of unsigned 16 bit int **/
-#define HIGH_BYTE_16(num) ((((uint16_t) num) >> 8) & 0xF)
+#define HIGH_BYTE_16(num) ((uint8_t) ((num >> 8) & 0xF))
 /** second 8 bits of unsigned 16 bit int **/
-#define LOW_BYTE_16(num)  (((uint16_t) num) & 0xFF)
+#define LOW_BYTE_16(num)  ((uint8_t) ((num & 0xFF)))
 /** unpack two 8 bit integers into a 16 bit integer **/
-#define UNPACK_16(a, b) ((uint16_t) ((((uint8_t) a) << 8) | ((uint8_t) b)))
+#define UNPACK_16(a, b) ((uint16_t) ((a << 8) | b))
 
 int32_t mod(int32_t x, int32_t m);
 float floatMod(float x, float m);
